fix repo create validation
